Improve grow_board strategy and increase search horizon.
In addition, remove some dead/commented out code and use
grow_board all the time in puzzle.cc.

The strategy involves the observation that if a board with only
some elements - those closest to the end square - is unsolvable,
then the full board is also unsolvable. This improves speed because
it's very hard to conclude that a board is unsolvable, much easier
to determine that it's solvable.