Add simple move ordering and horizon shortening to minmax.
The former improves performance considerably while the latter
may be more marginal, but it's free, so why not?

It's also pretty unsophisticated; I'd like to use the prior depth's
PV or transposition table to guide the next depth in IDDFS. But if
I do that, I either have to pull IDDFS into minmax or inelegantly
export the PV or transposition table from the latter to the former.

Neither choice is very appealing.