Add stub .ZZT writer program (doesn't work yet due to linux-reconstruction).
Also make grow_board output information about how deep one has to go
in a seemingly unsolvable puzzle's game tree before managing to solve
the puzzle in the worst case. (It currently seems to be at the max
of height and width of the board.)

It might be possible to use this to stop grow_board early without having
to test *every* board up to 30 or whatever MAX_DEPTH is.