Add empirical heuristic for refutations in grow_board.
It appears that if a tile is added, most of the time if a solution
isn't found at (xsize + ysize) steps deeper in the tree than the current
solution, there is no solution to be had. Sometimes this fails, but the
speedup we get in proving puzzles unsolvable is worth it IMHO.

Also boost the max depth to take advantage of this.