Add weighted tile generation and retries to generator.cc's grow_board.
Weighted tile generation lets the frequency of the different
obstacles be random per board; retries lets the solver try skipping
past a number of tiles if they make the board unsolvable, thus
increasing the chance that a solvable complex puzzle is found.

The latter may still be a bit too slow due to how the retries
mess up the IDDFS. Better code is needed (it would be better to
just pretend that tile didn't happen, but that would require some
extraction of functionality that I don't have the time to do right
now).